Former-commit-id: 54ef257d370c29905059eee5be2a374777a19948

+++ PLATFORM/Modules/DynamicExpression/VirtoCommerce.DynamicExpressionModule.Data/Promotion/Conditions/CartConditions/ConditionAtNumItemsInCart.cs
@@ -8,7 +8,7 @@ namespace VirtoCommerce.DynamicExpressionModule.Data.Promotion
 	//[] [] items are in shopping cart
 	public class ConditionAtNumItemsInCart : ConditionBase, IConditionExpression
 	{
-		public decimal NumItem { get; set; }
+		public int NumItem { get; set; }
 
 		public bool Exactly { get; set; }
